# docs/swagger.yaml
definitions:
  api.authUserRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.authUserResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session_id:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    type: object
  api.createUserRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.createWalletRequest:
    properties:
      currency:
        type: string
    required:
    - currency
    type: object
  api.userResponse:
    properties:
      username:
        type: string
    type: object
  arfcasesqlc.Wallet:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: do test
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: index sample
      tags:
      - example
  /auth:
    post:
      consumes:
      - application/json
      description: Auth user & return session token (jwt)
      parameters:
      - description: Auth user params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.authUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authUserResponse'
      summary: Auth user
      tags:
      - user
  /users:
    post:
      consumes:
      - application/json
      description: Create user with username & password
      parameters:
      - description: Create user params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.userResponse'
      summary: Create user
      tags:
      - user
  /wallets:
    get:
      consumes:
      - application/json
      description: List wallets of the user
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page Id
        in: query
        name: page_id
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/arfcasesqlc.Wallet'
            type: array
      summary: List wallets
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create wallet for the user
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create wallet params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/arfcasesqlc.Wallet'
      summary: Create wallet
      tags:
      - wallet
swagger: "2.0"